Create ArcaRewardClaimerV1, and use interfaces
It took me a while to understand that the contract size was not going down because I needed to use interfaces. Otherwise, if you just import the contract, it slurps up all the byte code and you end up with something bigger instead of smaller.

I added a dev dependency that helps with the generation of interfaces.

Also, we really need unit tests lol