Simplify pipeline
Here I removed the idea of having two jobs.
Although it would be nice to have a more intuitive separation of concern, in practice it makes things a little more complicated because we need to pass the contract artifacts to the UI job. Not impossible, mind you, just not necessary at the moment.