Testnet deployment push
Success getting Sonic Testnet USDC from S!
I was able to deploy also on the testnet an arca vault! TEST1-USDC.

We (Claude and I) discovered that the second token needs to be whitelisted. Only two options: wS or USDC.

There are many problems in the UI left, but good progress. I'm committing many scripts, but a lot of them can be cleaned up afterwards... it was for debugging.