Merge dev into main for reward fix (#33)
* Mention the act command in README

* Create deploy-shadow-vault-upgrade script

In case we need to migrate vaults due to a code fix

* Add sleep after deployment of contract

* Run npm audit fix

* Add MIT license for UI

* WIP 1

* Update UI to use new shadow vault with deposit fix

* Update ShadowStrategy reward harvesting

I realised that my code was assuming that we were working with IGauge, but it's actually IGaugeV3 that we need to use.

This commit is WIP for the Harvest Reward fix. I'm trying in this attempt to use the Voter.

* Add sleep for shadow strat deploy too

* Add more sleeps in script

* WIP testing different approaches...

not working yet (harvest reward or exit position)

* Create SHADOW_WITHDRAWAL_FIX.md

Even this proposal isn't perfect, because it seems that the vault isn't aware of the amount of money that it has, which causes funds to not be accessible. WIP

* Create basic test structure for previewShares

* Implement CLI recoverERC20

* Print raw vault balance in CLI tool

I was able to withdraw my user reward from a shadow pool by depositing a small amount, since that calls _modifyUser

* CLI: accept absent strategy due to emergency mode

* Upgrade hardhat.config.ts to use v2 API, deploy

Now it builds (the solidity)

Update .gitignore

WIP upgrade to hardhat v3

Upgrade hardhat.config.ts to use v2 API, deploy

Revert "WIP upgrade to hardhat v3"

This reverts commit 3baf1c27935047f7c84d1f215bdfc6e4c2f64559.

* Update .gitignore

* Delete foundry tests

* Update hardhat.config.ts

* Update claude settings

* Create vault-factory-cli.ts

* Added comments in OracleRewardShadowVault.sol

The comments were originally from the metropolis vaults, but they apply to here

* Refactor shadow strategy reward claiming

Instead of having a bunch of different for loops one after the other, combine them into one loop.
Also, I added the "sweep token" step even for the reward tokens, let's see if it works by the logs

* WIP make shadow reward system closer to metropolis

Not tested yet, but in theory this should help

* Deploy multi-dex with a shadow vault

wS-USDC Pool: 0x324963c267C354c7660Ce8CA3F5f167E05649970

Deployed:
Vault Address: 0x81897b30c38A14c8B28B9Ab30Daab6BF4D84b340
Strategy Address: 0x203a2caDff2C755E5eF7ffbE839e8EA35F25409F

* Add claimRewards() to shadow vault CLI

* Delete SHADOW_WITHDRAWAL_FIX.md

* Use gauge instead of NPM for shadow reward claiming

* Deploy shadow strat upgrade

* WIP use typechain in UI

Instead of relying on a manual ABI.

This means that I will need to modify the AWS pipeline to compile and generate the typechain.

Here I would like to remove the use of "any"

* Improve type safety for abi (remove any)

* Generate hardhat typechain in amplify

Revert "Generate hardhat typechain in amplify"

This reverts commit a5d532eef32dc0d8284ac4237492174d9cdb6bae.

No longer ignore typechain

Not ideal to commit a build result, but it's useful to have it present immediately for the AWS amplify pipeline

* Commit typechain-types